Fix values in array column and use correct labels (#211)
* add `heatmap` to renderer

* add property `labels` and assign `years`

in order to avoid labels `Column 0`, ...

* use `getRawNumbers()` instead of `getLabels`

and use `getNumberFormat()` instead of `DEFAULT_FORMATTER`

* set `raw` to true to get the correct numbers

* set `style.left` to 0 if `leftWhisker` is negative

* use `getRawValue()` instead of `getValues()`

in order to retrieve the correct values

* add the property `labels` to `BooleanColumn`

* use correct labels in `UpSetCellRenderer`

* use `getRawNumbers()` instead of `getNumbers()`

in order to retrieve the right values

* remove `=` from comparision

* adapt `whiskers.style.left`

if the summary visualization is rendered, the originally used
value is assigned, otherwise it is set to 0 in order to avoid
misplacing

* remove multiplication with `width` for `outliers`

in order to avoid displaying them outside the column

* revert changes for outliers

* adapt styling of outliers

if they are placed outside the column

* revert changes for outliers

because they affect the boxplot

* rename variable

Co-Authored-By: Holger Stitz <holger.stitz@datavisyn.io>

* rename  variable

* set default value for `raw` as `true`

instead of setting the flag in the function

* add boxplot example

* revert deprecated solution for boxplot calc

* split `renderDOMBoxPlot()'

- create `renderItemDOMBoxPlot()` and
`renderSummaryDOMBoxPlot()'
- add `scaleLinear` for `summaryRenderer' because the
calculations base on a range from 0 to 1 and without the
mapping, the wrong values are used

* Moved comments to corresponding datasets

* Correct some expected boxplot values

The calculated and visualized boxplot values by LineUp are matching with the expected. I checked the values using this page: https://goodcalculators.com/box-plot-maker/

* Fix NumberColumnBuilder for linear mapping

Before the `mapping()` returned `undefined` incase of a linear mapping. Now it returns the builder instance.

* Set boxplot renderer and add inverted mapping

* remove `renderSummaryDOMBoxPlot' and scaleLinear

fix problem with labels and inverted data by setting the raw
flag in `createSummary` and change the function according
to `createGroup`; use `mappedSummary` and `rawSummary`

* revert setting default value for `raw`

* remove unused `!` from variable `data.outlier`

* Revert some code and code formatting

* Revert to col.getLabels()

* Fix tslint

* Revert getRawNumbers() in HistogramCellRenderer

* Revert label in UpSetCellRenderer

The UpSet renderer is mainly used for categorical or set columns. Categorical arrays is a corner case. So it seems like that the category generation might be wrong but imo the UpSet renderer should be just based on categories.